Method finalizeDirs now prompts by default whether to delete temp dirs, unless new boolean field "force" is true.

+++ tools/DataDistributer.java
@@ -1,6 +1,9 @@
 package tools;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.HashMap;
@@ -26,6 +29,7 @@ public class DataDistributer {
 	String user;
 	String pwd;
 	boolean debug = false;
+	boolean force = false;
 	
 	/**
 	 * Construct DataDistributer object passing only source db. The destination db will be same as source.
@@ -60,6 +64,10 @@ public class DataDistributer {
 		this.debug=debug;
 	}
 	
+	public void setForce(boolean force){
+		this.force=force;
+	}
+	
 	public void setDumpDir(String dumpdir){
 		this.dumpdir=dumpdir;
 	}
@@ -87,9 +95,27 @@ public class DataDistributer {
 	    if (debug) {
 	    	System.out.println("Temporary directory "+dumpdir+" was not removed. You must remove it manually");
 	    } else {
-	    	System.out.println("Removing temporary directory "+dumpdir);
-	    	//TODO must capture exit state and print to error if problems deleting dir
-	    	SystemCmd.exec("rm -rf "+dumpdir);
+	    	if (force) {
+	    		System.out.println("Removing temporary directory "+dumpdir);
+	    		//TODO must capture exit state and print to error if problems deleting dir
+	    		SystemCmd.exec("rm -rf "+dumpdir);
+	    	} else {
+	    		System.out.println("Would you like to remove the temporary data directory '"+dumpdir+"' ? (y/n)");
+	    		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+	    		try {
+	    			String answer = br.readLine();
+	    			if (answer.equals("y") || answer.equals("yes") || answer.equals("Y") || answer.equals("YES") || answer.equals("Yes")) {
+	    	    		System.out.println("Removing temporary directory "+dumpdir);
+	    	    		SystemCmd.exec("rm -rf "+dumpdir);
+	    			} else {
+	    		    	System.out.println("Temporary directory "+dumpdir+" was not removed.");
+	    			}
+	    		} catch (IOException e) {
+	    			System.err.println("I/O error while reading user input.");
+	    			e.printStackTrace();
+	    			System.exit(2);
+	    		}
+	    	}
 	    }
 	}
